+++ src/Machine.Specifications.Runner.ReSharper.Adapters/ExceptionResultExtensions.cs
@@ -1,11 +1,12 @@
 ﻿using System.Collections.Generic;
 using System.Reflection;
+using Machine.Specifications.Runner.Utility;
 
 namespace Machine.Specifications.Runner.ReSharper.Adapters
 {
     public static class ExceptionResultExtensions
     {
-        public static IEnumerable<Utility.ExceptionResult> Flatten(this Utility.ExceptionResult result)
+        public static IEnumerable<ExceptionResult> Flatten(this ExceptionResult result)
         {
             var exception = result;
