+++ src/Machine.Specifications.Runner.ReSharper.Adapters/Machine.Specifications.Runner.ReSharper.Adapters.csproj
@@ -10,13 +10,20 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <!--<PackageReference Include="ILRepack" Version="2.0.18" PrivateAssets="All" />
-    <PackageReference Include="Machine.Specifications.Runner.Utility" Version="1.0.1-add-netstandard-uti1" PrivateAssets="All" />-->
-    <PackageReference Include="Machine.Specifications" Version="1.0.0" PrivateAssets="All" />
+    <PackageReference Include="ILRepack" Version="2.0.18" PrivateAssets="All" />
+    <PackageReference Include="Machine.Specifications.Runner.Utility" Version="1.0.1" PrivateAssets="All" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Machine.Specifications.Runner.ReSharper.Adapters.Tasks\Machine.Specifications.Runner.ReSharper.Adapters.Tasks.csproj" />
   </ItemGroup>
+
+  <Target Name="ILRepack" AfterTargets="Build" Condition="'$(TargetFramework)' != ''">
+    <PropertyGroup>
+      <ILRepackCommand>"$(ILRepack)" --internalize --norepackres --out:"$(OutputPath)$(AssemblyName).dll" "$(TargetPath)" "$(OutputPath)Machine.Specifications.Runner.Utility.dll"</ILRepackCommand>
+    </PropertyGroup>
+
+    <Exec Command="$(ILRepackCommand)" />
+  </Target>
 
 </Project>
