+++ src/Machine.Specifications.Runner.ReSharper.Adapters/TestRunListener.cs
@@ -4,16 +4,17 @@ using System.Linq;
 using JetBrains.ReSharper.TestRunner.Abstractions;
 using JetBrains.ReSharper.TestRunner.Abstractions.Objects;
 using Machine.Specifications.Runner.ReSharper.Adapters.Tasks;
+using Machine.Specifications.Runner.Utility;
 
 namespace Machine.Specifications.Runner.ReSharper.Adapters
 {
-    public class TestRunListener : Utility.ISpecificationRunListener
+    public class TestRunListener : ISpecificationRunListener
     {
         private readonly ITestExecutionSink server;
 
         private readonly TestContext context;
 
-        private Utility.ContextInfo currentContext;
+        private ContextInfo currentContext;
 
         private int specifications;
 
@@ -27,12 +28,12 @@ namespace Machine.Specifications.Runner.ReSharper.Adapters
             this.context = context;
         }
 
-        public void OnAssemblyStart(Utility.AssemblyInfo assemblyInfo)
+        public void OnAssemblyStart(AssemblyInfo assemblyInfo)
         {
             Environment.CurrentDirectory = GetWorkingDirectory(context.AssemblyLocation);
         }
 
-        public void OnAssemblyEnd(Utility.AssemblyInfo assemblyInfo)
+        public void OnAssemblyEnd(AssemblyInfo assemblyInfo)
         {
             Output(default, assemblyInfo.CapturedOutput);
         }
@@ -45,7 +46,7 @@ namespace Machine.Specifications.Runner.ReSharper.Adapters
         {
         }
 
-        public void OnContextStart(Utility.ContextInfo contextInfo)
+        public void OnContextStart(ContextInfo contextInfo)
         {
             specifications = 0;
             errors = 0;
@@ -61,7 +62,7 @@ namespace Machine.Specifications.Runner.ReSharper.Adapters
             }
         }
 
-        public void OnContextEnd(Utility.ContextInfo contextInfo)
+        public void OnContextEnd(ContextInfo contextInfo)
         {
             var result = TestResult.Inconclusive;
 
@@ -87,7 +88,7 @@ namespace Machine.Specifications.Runner.ReSharper.Adapters
             server.TestFinished(task, message, result);
         }
 
-        public void OnSpecificationStart(Utility.SpecificationInfo specificationInfo)
+        public void OnSpecificationStart(SpecificationInfo specificationInfo)
         {
             var task = context.GetBehaviorTask(currentContext, specificationInfo) ??
                        context.GetSpecificationTask(specificationInfo);
@@ -102,7 +103,7 @@ namespace Machine.Specifications.Runner.ReSharper.Adapters
             specifications += 1;
         }
 
-        public void OnSpecificationEnd(Utility.SpecificationInfo specificationInfo, Utility.Result result)
+        public void OnSpecificationEnd(SpecificationInfo specificationInfo, Result result)
         {
             var task = context.GetBehaviorTask(currentContext, specificationInfo) ??
                        context.GetSpecificationTask(specificationInfo);
@@ -114,30 +115,30 @@ namespace Machine.Specifications.Runner.ReSharper.Adapters
 
             Output(task, specificationInfo.CapturedOutput);
 
-            if (result.Status == Utility.Status.Failing)
+            if (result.Status == Status.Failing)
             {
                 errors++;
                 server.TestException(task, GetExceptions(result.Exception));
                 server.TestFinished(task, GetExceptionMessage(result.Exception), TestResult.Failed);
             }
-            else if (result.Status == Utility.Status.Passing)
+            else if (result.Status == Status.Passing)
             {
                 successes++;
                 server.TestFinished(task, string.Empty, TestResult.Success);
             }
-            else if (result.Status == Utility.Status.NotImplemented)
+            else if (result.Status == Status.NotImplemented)
             {
                 Output(task, "Not implemented");
                 server.TestFinished(task, "Not implemented", TestResult.Inconclusive);
             }
-            else if (result.Status == Utility.Status.Ignored)
+            else if (result.Status == Status.Ignored)
             {
                 Output(task, "Ignored");
                 server.TestFinished(task, string.Empty, TestResult.Ignored);
             }
         }
 
-        public void OnFatalError(Utility.ExceptionResult exceptionResult)
+        public void OnFatalError(ExceptionResult exceptionResult)
         {
             server.TestOutput(null, "Fatal error: " + exceptionResult.Message, TestOutputType.STDOUT);
             server.TestException(null, GetExceptions(exceptionResult));
@@ -154,14 +155,14 @@ namespace Machine.Specifications.Runner.ReSharper.Adapters
             }
         }
 
-        private ExceptionInfo[] GetExceptions(Utility.ExceptionResult result)
+        private ExceptionInfo[] GetExceptions(ExceptionResult result)
         {
             return result.Flatten()
                 .Select(x => new ExceptionInfo(x.FullTypeName, x.Message, x.StackTrace))
                 .ToArray();
         }
 
-        private string GetExceptionMessage(Utility.ExceptionResult result)
+        private string GetExceptionMessage(ExceptionResult result)
         {
             var exception = result.Flatten().FirstOrDefault();
