--- src/Machine.Specifications.Runner.ReSharper.Adapters/Utility/FullNames.cs
@@ -1,15 +1,0 @@
-﻿namespace Machine.Specifications.Runner.Utility
-{
-    public static class FullNames
-    {
-        public static readonly string ActDelegateAttribute = "Machine.Specifications.ActDelegateAttribute";
-        public static readonly string SetupDelegateAttribute = "Machine.Specifications.SetupDelegateAttribute";
-        public static readonly string BehaviorDelegateAttribute = "Machine.Specifications.BehaviorDelegateAttribute";
-        public static readonly string AssertDelegateAttribute = "Machine.Specifications.AssertDelegateAttribute";
-        public static readonly string CleanupDelegateAttribute = "Machine.Specifications.CleanupDelegateAttribute";
-        public static readonly string SubjectAttribute = "Machine.Specifications.SubjectAttribute";
-        public static readonly string TagsAttribute = "Machine.Specifications.TagsAttribute";
-        public static readonly string IgnoreAttribute = "Machine.Specifications.IgnoreAttribute";
-        public static readonly string BehaviorsAttribute = "Machine.Specifications.BehaviorsAttribute";
-    }
-}
