+++ src/Machine.Specifications.Runner.ReSharper/Machine.Specifications.Runner.ReSharper.csproj
@@ -24,7 +24,7 @@
 
   <ItemGroup>
     <PackageReference Include="JetBrains.ReSharper.SDK" Version="2020.2.1" PrivateAssets="All" />
-    <PackageReference Include="Machine.Specifications.Runner.Utility" Version="1.0.1-add-netstandard-uti1" PrivateAssets="All" />
+    <PackageReference Include="Machine.Specifications.Runner.Utility" Version="1.0.1" PrivateAssets="All" />
     <PackageReference Include="Wave" Version="[$(WaveVersion)]" />
   </ItemGroup>
 
@@ -44,11 +44,11 @@
       <TfmSpecificPackageFile Include="$(OutputPath)\Machine.Specifications.Runner.ReSharper.Runner.dll" PackagePath="DotFiles" />
       <TfmSpecificPackageFile Include="$(OutputPath)\Machine.Specifications.Runner.Utility.dll" PackagePath="DotFiles" />
 
-      <TfmSpecificPackageFile Include="$(SolutionDir)\src\Machine.Specifications.Runner.ReSharper.Adapters\bin\$(ConfigurationName)\net40\Machine.Specifications.Runner.ReSharper.Adapters.dll" PackagePath="DotFiles/TestRunner/Adapters/MSpec/net40" />
-      <TfmSpecificPackageFile Include="$(SolutionDir)\src\Machine.Specifications.Runner.ReSharper.Adapters.Tasks\bin\$(ConfigurationName)\net40\Machine.Specifications.Runner.ReSharper.Adapters.Tasks.dll" PackagePath="DotFiles/TestRunner/Adapters/MSpec/net40" />
+      <TfmSpecificPackageFile Include="$(SolutionDir)\src\Machine.Specifications.Runner.ReSharper.Adapters\bin\$(ConfigurationName)\net40\Machine.Specifications.Runner.ReSharper.Adapters.net40.dll" PackagePath="DotFiles/TestRunner/Adapters/MSpec/net40" />
+      <TfmSpecificPackageFile Include="$(SolutionDir)\src\Machine.Specifications.Runner.ReSharper.Adapters.Tasks\bin\$(ConfigurationName)\net40\Machine.Specifications.Runner.ReSharper.Adapters.Tasks.net40.dll" PackagePath="DotFiles/TestRunner/Adapters/MSpec/net40" />
 
-      <TfmSpecificPackageFile Include="$(SolutionDir)\src\Machine.Specifications.Runner.ReSharper.Adapters\bin\$(ConfigurationName)\netstandard2.0\Machine.Specifications.Runner.ReSharper.Adapters.dll" PackagePath="DotFiles/TestRunner/Adapters/MSpec/netstandard2.0" />
-      <TfmSpecificPackageFile Include="$(SolutionDir)\src\Machine.Specifications.Runner.ReSharper.Adapters.Tasks\bin\$(ConfigurationName)\netstandard2.0\Machine.Specifications.Runner.ReSharper.Adapters.Tasks.dll" PackagePath="DotFiles/TestRunner/Adapters/MSpec/netstandard2.0" />
+      <TfmSpecificPackageFile Include="$(SolutionDir)\src\Machine.Specifications.Runner.ReSharper.Adapters\bin\$(ConfigurationName)\netstandard2.0\Machine.Specifications.Runner.ReSharper.Adapters.netstandard20.dll" PackagePath="DotFiles/TestRunner/Adapters/MSpec/netstandard2.0" />
+      <TfmSpecificPackageFile Include="$(SolutionDir)\src\Machine.Specifications.Runner.ReSharper.Adapters.Tasks\bin\$(ConfigurationName)\netstandard2.0\Machine.Specifications.Runner.ReSharper.Adapters.Tasks.netstandard20.dll" PackagePath="DotFiles/TestRunner/Adapters/MSpec/netstandard2.0" />
     </ItemGroup>
   </Target>
